pyproject.toml
[tool.poetry]
name = "tarot"
version = "0.1.0"
description = ""
authors = ["Ian MacDonald <ian@imlabs.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
